craft_requirements.txt
torch==2.3.1
torchvision==0.2.1
pillow==6.1.0
opencv-python==4.10.0
scikit-image==0.14.2
tqdm==4.66.4